Minor refactoring to abstract unnecessary details.
